Add new endpoint verification with restricted token

+++ pkg/gateway/services/token.go
@@ -12,6 +12,15 @@ import (
 
 func (gws *GatewayService) ListTokens(ctx context.Context, req *pb.ListTokensRequest) (*pb.ListTokensResponse, error) {
 	authInfo, _ := auth.AuthInfoFromContext(ctx)
+
+	if auth.CheckWorkspaceRestrictedToken(authInfo) {
+		return &pb.ListTokensResponse{
+			Tokens: []*pb.Token{},
+			Ok:     false,
+			ErrMsg: "Access denied for workspace restricted tokens",
+		}, nil
+	}
+
 	workspaceId := authInfo.Workspace.Id
 	tokens, err := gws.backendRepo.ListTokens(ctx, workspaceId)
 	if err != nil {
@@ -51,6 +60,13 @@ func (gws *GatewayService) ListTokens(ctx context.Context, req *pb.ListTokensReq
 
 func (gws *GatewayService) CreateToken(ctx context.Context, req *pb.CreateTokenRequest) (*pb.CreateTokenResponse, error) {
 	authInfo, _ := auth.AuthInfoFromContext(ctx)
+
+	if auth.CheckWorkspaceRestrictedToken(authInfo) {
+		return &pb.CreateTokenResponse{
+			Ok:     false,
+			ErrMsg: "Access denied for workspace restricted tokens",
+		}, nil
+	}
 
 	tokenType := req.TokenType
 	if tokenType == "" {
@@ -95,6 +111,14 @@ func (gws *GatewayService) CreateToken(ctx context.Context, req *pb.CreateTokenR
 
 func (gws *GatewayService) ToggleToken(ctx context.Context, req *pb.ToggleTokenRequest) (*pb.ToggleTokenResponse, error) {
 	authInfo, _ := auth.AuthInfoFromContext(ctx)
+
+	if auth.CheckWorkspaceRestrictedToken(authInfo) {
+		return &pb.ToggleTokenResponse{
+			Ok:     false,
+			ErrMsg: "Access denied for workspace restricted tokens",
+		}, nil
+	}
+
 	token, err := gws.backendRepo.ToggleToken(ctx, authInfo.Workspace.Id, req.TokenId)
 	if err != nil {
 		return &pb.ToggleTokenResponse{
@@ -124,6 +148,13 @@ func (gws *GatewayService) ToggleToken(ctx context.Context, req *pb.ToggleTokenR
 
 func (gws *GatewayService) DeleteToken(ctx context.Context, req *pb.DeleteTokenRequest) (*pb.DeleteTokenResponse, error) {
 	authInfo, _ := auth.AuthInfoFromContext(ctx)
+
+	if auth.CheckWorkspaceRestrictedToken(authInfo) {
+		return &pb.DeleteTokenResponse{
+			Ok:     false,
+			ErrMsg: "Access denied for workspace restricted tokens",
+		}, nil
+	}
 	err := gws.backendRepo.DeleteToken(ctx, authInfo.Workspace.Id, req.TokenId)
 	if err != nil {
 		return &pb.DeleteTokenResponse{
